shape - 10

+++ geometry_shape/step_8/src/main/java/org/javaguru/geometry/shapes/Rectangle.java
@@ -2,10 +2,10 @@ package org.javaguru.geometry.shapes;
 
 class Rectangle extends Shape {
 
-    private int a;
-    private int b;
+    private double a;
+    private double b;
 
-    public Rectangle(String title, int a, int b) {
+    public Rectangle(String title, double a, double b) {
         super(title);
         this.a = a;
         this.b = b;
